__future__ contains the future features of Python that can be used in the current version.
Importing __future__ helps us write a code such that it is compatible with the current version and the future version of Python, so that the programmer need not worry about altering the code once the version switch has taken place.
It directs the compiler to compile a particular module using syntax or semantics that will be available in a specified future release of Python.
For more info : http://www.learningpython.com/2010/09/22/using-future-statements-to-prepare-for-the-future/